Move `SDL_SysWMmsg` struct to separate file
